requirements.txt
streamlit
scikit-learn==1.6.1
pandas
numpy
joblib
pdf2image
google-generativeai
python-dotenv
poppler-utils
